# docs/swagger.yaml
basePath: /v1
definitions:
  httpx.DataJsonResponse:
    properties:
      data: {}
    type: object
  httpx.ErrorJsonResponse:
    properties:
      error:
        type: string
    type: object
  httpx.MessageJsonResponse:
    properties:
      message:
        type: string
    type: object
...
  schemas.Admin:
    properties:
      email:
        type: string
      id:
        type: string
      password:
        type: string
      phone_number:
        type: string
      username:
        type: string
    type: object
...
  schemas.Alternative:
    properties:
      alternative:
        type: string
      description:
        type: string
      id:
        type: integer
    type: object
...
  schemas.LoginRequest:
    properties:
      password:
        type: string
      type:
        type: string
      username:
        type: string
    type: object
  schemas.LoginResponse:
    properties:
      id:
        type: string
      role:
        type: string
      school_id:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
...
  schemas.School:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  schemas.Statistic:
    properties:
      consistency_avg:
        type: number
      questionnare_completed:
        type: integer
      recommendation_acceptance:
        type: number
      registered_students:
        type: integer
    type: object
...
  schemas.UpdateAdminRequest:
    properties:
      email:
        type: string
      phone_number:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a Double Track REST API
  title: Double Track API
  version: "1.0"
paths:
  /admin/profile:
    get:
      consumes:
      - application/json
      description: Get admin profile
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/schemas.Admin'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Get admin profile
      tags:
      - Admin
    patch:
      consumes:
      - application/json
      description: Update admin profile
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update admin profile request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpx.MessageJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Update admin profile
      tags:
      - Admin
  /alternatives:
    get:
      consumes:
      - application/json
      description: Get all alternatives
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schemas.Alternative'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Get alternatives
      tags:
      - Common
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login authentication for student and admin
      parameters:
      - description: Login request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schemas.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/schemas.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Login authentication
      tags:
      - Auth
...
  /school:
    get:
      consumes:
      - application/json
      description: Get current authenticated user shcool
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/schemas.School'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Get school
      tags:
      - Common
      - Student
      - Admin
  /statistics:
    get:
      consumes:
      - application/json
      description: Get statistics
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/schemas.Statistic'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Get statistic
      tags:
      - Common
      - Admin
...
swagger: "2.0"